+++ .gitignore
@@ -13,3 +13,7 @@ CLAUDE.md
 
 # Environment variables
 .env
+
+old_test
+
+old_Src
