+++ src/lib.rs
@@ -1,5 +1,2 @@
-pub mod utils;
-pub mod notes;
-pub mod deploy;
-pub mod config;
-pub mod scripts;
+pub mod storage;
+pub mod domain;
