+++ tests/test_utils.rs
@@ -1,0 +1,158 @@
+use std::{fs, path::Path, sync::Arc};
+
+use anyhow::Ok;
+use miden_assembly::{Assembler, DefaultSourceManager, Library, LibraryPath, ast::{Module, ModuleKind}};
+use miden_client::{ScriptBuilder, account::{Account, AccountBuilder, AccountId, AccountStorageMode}, asset::FungibleAsset, note::{Note, NoteAssets, NoteExecutionHint, NoteInputs, NoteMetadata, NoteRecipient, NoteTag, NoteType}, testing::account_id::ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1, transaction::OutputNote};
+use miden_crypto::{Felt, Word};
+use miden_lib::{account::auth, transaction::TransactionKernel};
+use miden_objects::account::AccountComponent;
+use miden_testing::{Auth, MockChain, TransactionContextBuilder};
+use midenname_contracts::storage::naming_storage;
+use rand::{Rng, SeedableRng};
+use rand_chacha::ChaCha20Rng;
+
+fn create_test_naming_account() -> Account {
+    let storage_slots = naming_storage();
+    let code = fs::read_to_string(Path::new("./masm/accounts/naming.masm")).unwrap();
+
+    let component = AccountComponent::compile(
+        code.clone(), 
+        TransactionKernel::assembler().with_debug_mode(true), 
+        storage_slots
+    ).unwrap().with_supports_all_types();
+
+    let account = AccountBuilder::new(ChaCha20Rng::from_os_rng().random())
+        .with_auth_component(auth::NoAuth)
+        .with_component(component)
+        .storage_mode(AccountStorageMode::Public)
+        .build_existing().unwrap();
+
+    account
+}
+
+pub fn get_test_prices() -> Vec<Felt> {
+    vec![Felt::new(0), Felt::new(123123), Felt::new(45645), Felt::new(789), Felt::new(555), Felt::new(123)]
+}
+
+pub struct TestingContext {
+    pub chain: MockChain,
+    pub owner: Account,
+    pub registrar_1: Account,
+    pub registrar_2: Account,
+    pub registrar_3: Account,
+    pub naming: Account,
+    pub fungible_asset: FungibleAsset,
+    pub one_year: u32
+}
+
+pub async fn init_naming() -> anyhow::Result<TestingContext> {
+    let mut builder = MockChain::builder();
+    let fungible_asset_1 = FungibleAsset::new(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into().unwrap(), 100000).unwrap();
+    let fungible_asset_2 = FungibleAsset::new(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into().unwrap(), 50000).unwrap();
+    let fungible_asset_3 = FungibleAsset::new(ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET_1.try_into().unwrap(), 20000).unwrap();
+
+    let owner_account = builder.add_existing_wallet(Auth::BasicAuth)?;
+    let domain_registrar_account = builder.add_existing_wallet_with_assets(Auth::BasicAuth, vec![fungible_asset_1.into()])?;
+    let domain_registrar_account_2 = builder.add_existing_wallet_with_assets(Auth::BasicAuth, vec![fungible_asset_2.into()])?;
+    let domain_registrar_account_3 = builder.add_existing_wallet_with_assets(Auth::BasicAuth, vec![fungible_asset_3.into()])?;
+    let naming_account = create_test_naming_account();
+    builder.add_account(naming_account.clone())?;
+    let mut mockchain = builder.build()?;
+    let one_year_time: u32 = 500;
+
+    println!("\nNaming Account: {}", naming_account.id().to_hex());
+    println!("Fungible Asset: {}", fungible_asset_1.faucet_id().to_hex());
+    println!("Owner Account: {}", owner_account.id().to_hex());
+
+    let initialize_inputs =NoteInputs::new([
+        Felt::new(owner_account.id().suffix().into()),
+        Felt::new(owner_account.id().prefix().into()),
+        Felt::new(0),
+        Felt::new(0),
+        Felt::new(one_year_time.into()),
+        Felt::new(0),
+        Felt::new(0),
+        Felt::new(0),
+    ].to_vec())?;
+    let init_note = create_note_for_naming("initialize_naming".to_string(), initialize_inputs, owner_account.id(), naming_account.id(), NoteAssets::new(vec![]).unwrap()).await?;
+    
+    let updated_naming_account = execute_note(&mut mockchain, init_note, naming_account.clone()).await?;
+    println!("Naming initialized");
+    // Set prices
+
+    let updated_naming_account = set_test_prices(&mut mockchain, owner_account.id(), updated_naming_account, fungible_asset_1.faucet_id()).await?;
+
+    Ok(TestingContext { chain: mockchain, owner: owner_account, registrar_1: domain_registrar_account, 
+        registrar_2: domain_registrar_account_2, registrar_3: domain_registrar_account_3, naming: updated_naming_account, 
+        fungible_asset: fungible_asset_1, one_year: one_year_time })
+}
+
+async fn execute_note(chain: &mut MockChain, note: Note, target: Account) -> anyhow::Result<Account> {
+    chain.add_pending_note(OutputNote::Full(note.clone()));
+    chain.prove_next_block()?;
+
+    let note_inputs = chain.get_transaction_inputs(target.clone(), None, &[note.id()], &[])?;
+    let tx_context = TransactionContextBuilder::new(target.clone())
+        .account_seed(None).tx_inputs(note_inputs).build()?;
+
+    let executed_tx = tx_context.execute().await?;;
+    let updated_account = chain.add_pending_executed_transaction(&executed_tx)?;
+    chain.prove_next_block()?;
+
+    Ok(updated_account)
+}
+
+async fn create_note_for_naming(name: String, inputs: NoteInputs, sender: AccountId, target_id: AccountId, assets: NoteAssets) -> anyhow::Result<Note> {
+    let note_code = fs::read_to_string(Path::new(&format!("./masm/notes/{}.masm", name)))?;
+    let naming_code = fs::read_to_string(Path::new("./masm/accounts/naming.masm")).unwrap();
+    let library = create_library(naming_code, "miden_name::naming")?;
+
+    let note_script = ScriptBuilder::new(true)
+        .with_dynamically_linked_library(&library)
+        .unwrap()
+        .compile_note_script(note_code)
+        .unwrap();
+
+    let recipient = NoteRecipient::new(Word::default(), note_script, inputs.clone());
+    let tag = NoteTag::from_account_id(target_id);
+    let metadata = NoteMetadata::new(sender, NoteType::Public, tag, NoteExecutionHint::Always, Felt::new(0))?;
+    let note = Note::new(assets, metadata, recipient);
+    Ok(note)
+}
+
+fn create_library(account_code: String, library_path: &str) -> anyhow::Result<Library> {
+    let assembler: Assembler = TransactionKernel::assembler().with_debug_mode(true);
+    let source_manager = Arc::new(DefaultSourceManager::default());
+    let module = Module::parser(ModuleKind::Library).parse_str(
+        LibraryPath::new(library_path)?,
+        account_code,
+        &source_manager,
+    ).unwrap();
+    let library = assembler.clone().assemble_library([module]).unwrap();
+
+    Ok(library)
+}
+
+async fn set_test_prices(chain: &mut MockChain, tx_sender: AccountId, naming: Account, asset: AccountId) -> anyhow::Result<Account> {
+    let prices = get_test_prices();
+    let mut updated_naming_account = naming.clone();
+    for i in 1..=5 {
+        println!("\nSetting price for letter {}", i);
+        let one_letter_inputs = NoteInputs::new([
+            prices[i as usize],
+            Felt::new(0),
+            Felt::new(0),
+            Felt::new(0),
+            Felt::new(asset.suffix().into()),
+            Felt::new(asset.prefix().into()),
+            Felt::new(i as u64), // letter_count
+            Felt::new(0),
+        ].to_vec())?;
+
+        let one_letter_note = create_note_for_naming("set_price".to_string(), one_letter_inputs, tx_sender, naming.id(), NoteAssets::new(vec![]).unwrap()).await?;
+
+        updated_naming_account = execute_note(chain, one_letter_note, updated_naming_account.clone()).await?;
+    }
+
+    Ok(updated_naming_account)
+}
